LFU-416 imx: cmd_dek: Fix build warning in blob_encap_dek
Fix below build warning introduced by commit bf07f51
(LFOPTEE-177 imx: cmd_dek: add ELE DEK Blob generation support)

In file included from ./arch/arm/include/asm/io.h:341,
                 from include/fsl_sec.h:13,
                 from arch/arm/mach-imx/cmd_dek.c:16:
include/cpu_func.h:68:39: note: expected ‘long unsigned int’
but argument is of type ‘u8 *’ {aka ‘unsigned char *’}
   68 | void flush_dcache_range(unsigned long start, unsigned long stop);

+++ arch/arm/mach-imx/cmd_dek.c
@@ -353,7 +353,7 @@ static int blob_encap_dek(u32 src_addr, u32 dst_addr, u32 len)
 	
 	/* Flush the cache */
 	flush_dcache_range(src_addr, src_addr + in_size);
-	flush_dcache_range(dst_ptr, (ulong)(dst_ptr +
+	flush_dcache_range((ulong)dst_ptr, (ulong)(dst_ptr +
 			roundup(out_size, ARCH_DMA_MINALIGN)));
 
 	/* Call ELE */
